removed postinstall script EF-2420
