+++ package.json
@@ -4,7 +4,7 @@
     "type": "git",
     "url": "git://github.com/roXtra/processhub-sdk.git"
   },
-  "version": "9.33.0-5",
+  "version": "9.33.0-8",
   "author": {
     "name": "ProcessHub",
     "email": "info@processhub.com",
@@ -52,9 +52,9 @@
     "xmlhttprequest": "^1.8.0"
   },
   "scripts": {
-    "build": "rimraf ./lib && tsc && npm run copytypings",
-    "test": "npm run build && npm run lint && npm run testcoverage",
-    "lint": "eslint --ext .ts,.tsx ./src",
+    "build": "npm run copy-kendo-headers && rimraf ./lib && tsc && npm run copytypings",
+    "test": "npm run copy-kendo-headers && npm run build && npm run lint && npm run testcoverage",
+    "lint": "npm run copy-kendo-headers && eslint --ext .ts,.tsx ./src",
     "testall": "ts-mocha ./src/**/*.test.ts --exit --timeout 10000 --recursive",
     "testcoverage": "nyc --reporter=lcov --reporter=text-summary npm run testall",
     "copytypings": "copyfiles -u 1 \"src/**/*.d.ts\" lib && copyfiles -u 1 \"src/*.d.ts\" lib && copyfiles -u 1 src/process/types/index.d.ts lib",
@@ -62,8 +62,7 @@
     "release-bump-preview-version": "npm run check:format && npm run lint && npm version prerelease",
     "check:format": "prettier --config .prettierrc -c ./src",
     "format": "prettier --config .prettierrc -w ./src",
-    "postinstall": "npm run copy-kendo-headers",
-    "copy-kendo-headers": "copyfiles -u 2 ./templates/kendo-headers/* node_modules/@progress/kendo-react-data-tools/dist/npm/header"
+    "copy-kendo-headers": "cp ./src/templates/kendo-headers/* ./node_modules/@progress/kendo-react-data-tools/dist/npm/header || copyfiles -u 3 ./src/templates/kendo-headers/* node_modules/@progress/kendo-react-data-tools/dist/npm/header"
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
